+++ kyuubi-common/src/main/scala/org/apache/kyuubi/config/KyuubiConf.scala
@@ -1462,7 +1462,7 @@ object KyuubiConf {
       .createOptional
 
   val SESSION_PROGRESS_ENABLE: ConfigEntry[Boolean] =
-    buildConf("kyuubi.session.progress.enable")
+    buildConf("kyuubi.operation.progress.enabled")
       .doc("Whether to enable the operation progress. When true," +
         " the operation progress will be returned in `GetOperationStatus`.")
       .version("1.6.0")
